created update and get specific listing api routes

+++ server/controller/listing.js
@@ -40,10 +40,53 @@ export const deleteListing = async (req, res, next) => {
       return next("404", "Listing not found!");
     }
 
+    if (req.user.id !== checkListing.userRef) {
+      return next(
+        errorHandler("401", "You are not authorized to delete the listing")
+      );
+    }
+
     await LISTING.findByIdAndDelete(req.params.id);
     res.status(200).json({
       success: true,
       message: "Listing deleted successfully",
+    });
+  } catch (e) {
+    next(e);
+  }
+};
+
+export const getSpecificListing = async (req, res, next) => {
+  try {
+    const listing = await LISTING.findById(req.params.id);
+    if (!listing) {
+      return next(errorHandler("404", "Listing not found!"));
+    }
+
+    res.status(200).json(listing);
+  } catch (e) {
+    next(e);
+  }
+};
+
+export const updateListing = async (req, res, next) => {
+  try {
+    const listing = await LISTING.findById(req.params.id);
+    if (!listing) {
+      return next(errorHandler("404", "Listing not found!"));
+    }
+
+    if (req.user.id !== listing.userRef) {
+      return next(
+        errorHandler("401", "You are not authorized to update the listing")
+      );
+    }
+
+    await LISTING.findByIdAndUpdate(req.params.id, req.body, { new: true });
+
+    res.status(200).json({
+      success: true,
+      message: "Listing updated successfully",
     });
   } catch (e) {
     next(e);
